Remove animate() method from ROADMAP
I initially planned to add this method in the CSS plugin,
using emile.js [1] as implementation.

The rationale for this method was to provide visual feedback
to the user, e.g. when an element is clicked.

However, emile.js has not been under active development for some years
now, and this may be related to the fact that this kind of effect is now
in the realm of CSS3 transitions [2].

No JavaScript implementation is needed anymore, except maybe as a fallback
for older browsers. Since the visual feedback can be offered as a progressive
enhancement, and simply not implemented when not supported, I now see no need
to add this method in the framework.

[1]: https://github.com/madrobby/emile
[2]: http://www.w3.org/TR/css3-transitions/